requirements.txt
django==2.2.4
pillow
django-crispy-forms
pandas